Title Theoretical C and D matrix under assumption of Gaussianity

Title Theoretical C and D matrix under assumption of Gaussianity

Usage

C.gen(xm, ym, xs, corr = 0.5, n.iter = 1000)

Arguments

xm

Mean Matrix of Placebo

ym

Mean Matrix of Treatment

xs

Standard Deviation matrix

corr

Correlation between outcomes.

n.iter

Number of iterations

Value

C: Theoretical C under Gaussianity Assumption, D: Theoretical D under Gaussianity assumption

Examples

xm = -t(cbind(c(0, 0.738, 1.313, 3.109, 4.525, 5.864, 7.338),
          c(0, 0.668, 3.975, 5.931, 8.259, 11.989, 13.976)))
xs = t(cbind(c(0, 4.79, 5.43, 6.54, 7.37, 8.15, 9.11),
          c(0, 10.27, 12.85, 14.95, 15.35, 16.87, 18.19)))
alpha=1
beta=1
effect_size_1 = 2.21*alpha/6
effect_size_2 = 5.38*beta/6
ym = xm + t(cbind(c(0,(rep(effect_size_1,6))), c(0,(rep(effect_size_2,6)))))
C.gen(xm, ym, xs, n.iter=10)
#> $C
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.3209077 0.3191373 0.3202787 0.3224821 0.3198759 0.3222927
#> [2,] 0.3203815 0.3197951 0.3196591 0.3224807 0.3224591 0.3213256
#> [3,] 0.3200584 0.3203148 0.3203218 0.3218043 0.3216900 0.3223585
#> [4,] 0.3205726 0.3209949 0.3206170 0.3212945 0.3229912 0.3206036
#> [5,] 0.3206823 0.3216464 0.3195295 0.3204054 0.3224209 0.3197607
#> [6,] 0.3210736 0.3199761 0.3208121 0.3220017 0.3211203 0.3191869
#> 
#> $D
#>           [,1]      [,2]      [,3]      [,4]      [,5]      [,6]
#> [1,] 0.3191752 0.3202531 0.3219140 0.3203281 0.3214821 0.3214983
#> [2,] 0.3199125 0.3187930 0.3204907 0.3203946 0.3211722 0.3192914
#> [3,] 0.3206190 0.3192049 0.3192646 0.3213533 0.3220137 0.3220817
#> [4,] 0.3207663 0.3225235 0.3198931 0.3215526 0.3205296 0.3214093
#> [5,] 0.3220007 0.3216352 0.3213194 0.3191158 0.3211502 0.3219256
#> [6,] 0.3224811 0.3226684 0.3202296 0.3224309 0.3226069 0.3211046
#>