Title Estimate of Power from the data

Title Estimate of Power from the data

Usage

estimated_Power(alpha = 0.05, X_c, Y_c)

Arguments

alpha

Size of Test

X_c

Longitudinal Data for placebo of dimension (nx, K, T), where nx is the number of patients in the placebo group, K is the number of outcomes, and T is the number of time points

Y_c

Longitudinal Data for Treatment of dimension (ny, K, T), where ny is the number of patients in the treatment group, K is the number of outcomes, and T is the number of time points

Value

Estimated Power from the data

Examples

library(MASS)
xm = -t(cbind(c(0, 0.738, 1.313, 3.109, 4.525, 5.864, 7.338),
          c(0, 0.668, 3.975, 5.931, 8.259, 11.989, 13.976)))
xs = t(cbind(c(0, 4.79, 5.43, 6.54, 7.37, 8.15, 9.11),
          c(0, 10.27, 12.85, 14.95, 15.35, 16.87, 18.19)))
N = 500 # Total Number of Patients
nx = 2/5*N # Niumber of patients in Placebo
ny = 3/5*N # Number of patients in Treatment
K = 2 # Number of Outcomes
T = 6 # Number of visits other than baseline
sigma_u <- c(3.477, 7.232) # Random Effect Standard Deviation for Outcomes
corr <- matrix(c(1,0.5, 0.5, 1),2,2)  # Correlation between the outcomes
sigma = diag(sigma_u) %*% corr %*% diag(sigma_u)
alpha=1
beta=1
effect_size_1 = 2.21*alpha/6
effect_size_2 = 5.38*beta/6
ym = xm + t(cbind(c(0,(rep(effect_size_1,6))), c(0,(rep(effect_size_2,6)))))
X_placebo = array(NA, dim = c(nx, 2, 7))
X_dose_1 = array(NA, dim = c(ny, 2, 7))

for(i in 1:nx){
  X_placebo[i,,] = gen.fun.dep(xm, xs, corr, sigma)
}
for(i in 1:ny){
  X_dose_1[i,,] = gen.fun.dep(ym, xs, corr, sigma)
}
X_c = X_placebo
Y_c = X_dose_1
C.gen.hat(X_c, Y_c)
#> $C.hat
#>      [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#> [1,]    0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,]    0 0.1858960 0.1565563 0.1450974 0.1370631 0.1340145 0.1273275
#> [3,]    0 0.1565563 0.1946315 0.1753889 0.1672889 0.1578682 0.1485454
#> [4,]    0 0.1450974 0.1753889 0.1963280 0.1832724 0.1755310 0.1654889
#> [5,]    0 0.1370631 0.1672889 0.1832724 0.2013738 0.1904122 0.1817732
#> [6,]    0 0.1340145 0.1578682 0.1755310 0.1904122 0.2099745 0.1980152
#> [7,]    0 0.1273275 0.1485454 0.1654889 0.1817732 0.1980152 0.2098051
#> 
#> $D.hat
#>      [,1]      [,2]      [,3]      [,4]      [,5]      [,6]      [,7]
#> [1,]    0 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> [2,]    0 0.1900142 0.1634928 0.1514841 0.1455959 0.1349232 0.1292198
#> [3,]    0 0.1634928 0.2137997 0.1942236 0.1885475 0.1732515 0.1631944
#> [4,]    0 0.1514841 0.1942236 0.2257035 0.2090158 0.1917100 0.1783396
#> [5,]    0 0.1455959 0.1885475 0.2090158 0.2278281 0.2085668 0.1974241
#> [6,]    0 0.1349232 0.1732515 0.1917100 0.2085668 0.2175578 0.2041294
#> [7,]    0 0.1292198 0.1631944 0.1783396 0.1974241 0.2041294 0.2121668
#> 
estimated_Power(0.05, X_c, Y_c)
#> [1] 0.7815231