Title nparLD Test of Efficacy

Title nparLD Test of Efficacy

npld(X_c, Y_c)

Arguments

X_c

Longitudinal Data for placebo of dimension \((n_x, K, T)\), where \(n_x\) is the number of patients in the placebo gorup, K is the number of outcomes, and T is the number of time points

Y_c

Longitudinal Data for Treatment of dimension \((n_y, K, T)\), where \(n_y\) is the number of patients in the treatment gorup, K is the number of outcomes, and T is the number of time points

Value

A vector of p-values, length being same as the number of outcomes

References

Noguchi, K., Gel, Y. R., Brunner, E., & Konietschke, F. (2012). nparLD: an R software package for the nonparametric analysis of longitudinal data in factorial experiments. Journal of Statistical software, 50, 1-23.

Examples

library(MASS)
library(nparLD)
#> Warning: package 'nparLD' was built under R version 4.4.2
xm = -t(cbind(c(0, 0.738, 1.313, 3.109, 4.525, 5.864, 7.338),
          c(0, 0.668, 3.975, 5.931, 8.259, 11.989, 13.976)))
xs = t(cbind(c(0, 4.79, 5.43, 6.54, 7.37, 8.15, 9.11),
          c(0, 10.27, 12.85, 14.95, 15.35, 16.87, 18.19)))
N = 100 # Total Number of Patients
nx = 2/5*N # Niumber of patients in Placebo
ny = 3/5*N # Number of patients in Treatment
K = 2 # Number of Outcomes
T = 6 # Number of visits other than baseline
sigma_u <- c(3.477, 7.232) # Random Effect Standard Deviation for Outcomes
corr <- matrix(c(1,0.5, 0.5, 1),2,2)  # Correlation between the outcomes
sigma = diag(sigma_u) %*% corr %*% diag(sigma_u)
alpha=1
beta=1
effect_size_1 = 2.21*alpha/6
effect_size_2 = 5.38*beta/6
ym = xm + t(cbind(c(0,(rep(effect_size_1,6))), c(0,(rep(effect_size_2,6)))))
X_placebo = array(NA, dim = c(nx, 2, 7))
X_dose_1 = array(NA, dim = c(ny, 2, 7))

for(i in 1:nx){
  X_placebo[i,,] = gen.fun.dep(xm, xs, corr, sigma)
}
for(i in 1:ny){
  X_dose_1[i,,] = gen.fun.dep(ym, xs, corr, sigma)
}
X_c = X_placebo
Y_c = X_dose_1
npld(X_c, Y_c)
#>  F1 LD F1 Model 
#>  ----------------------- 
#>  Check that the order of the time and group levels are correct.
#>  Time level:   0 1 2 3 4 5 6 
#>  Group level:   0 1 
#>  If the order is not correct, specify the correct order in time.order or group.order.
#> 
#> 
#>  Warning(s):
#>  The covariance matrix is singular. 
#>  F1 LD F1 Model 
#>  ----------------------- 
#>  Check that the order of the time and group levels are correct.
#>  Time level:   0 1 2 3 4 5 6 
#>  Group level:   0 1 
#>  If the order is not correct, specify the correct order in time.order or group.order.
#> 
#> 
#>  Warning(s):
#>  The covariance matrix is singular. 
#> [1] 0.06541633 0.23247954