Title Theoretical Power when C and D are known

Title Theoretical Power when C and D are known

theoretical_power(alpha = 0.05, effect_size, nx, ny, K, T, C, D)

Arguments

alpha

Size of the Test

effect_size

Theoretical Effect Size

nx

Number of patients in the plcabo group

ny

Number of patients in the treatment group

K

Number of outcomes

T

Number of Time Points

C

Theoretical C matrix

D

Theoretica; D matrix

Value

Theoretical Power

References

Ghosh, D., Xu, X., Luo, S. (2025). Power and sample size calculation for multivariate longitudinal trials using the longitudinal rank sum test. arXiv preprint arXiv:2502.07152.

Examples

N = 500 # Total Number of Patients
nx = 2/5*N # Niumber of patients in Placebo
ny = 3/5*N # Number of patients in Treatment
K = 2 # Number of Outcomes
T = 6 # Number of visits other than baseline
C = matrix(rnorm(T^2), nrow = T)
D = matrix(rnorm(T^2), nrow = T)
theoretical_power(alpha = 0.05, effect_size = 1, nx, ny, K, T+1, C, D)
#>      [,1]
#> [1,]    1